requirements.txt
flask==2.3.3
python-dotenv==1.0.0
langchain==0.2.0
langchain-community==0.2.0
langchain-groq==0.1.0
langchain-pinecone==0.1.0
pinecone-client[grpc]==3.0.0
sentence-transformers==2.5.0
langchain-huggingface==0.1.0
huggingface_hub==0.20.0
pypdf==4.0.0